Modularized all functions and data into their own js files, exported them from those new js files, and imported them into all of the js files that they needed to be in
