+++ utils/renderHTML.js
@@ -1,0 +1,79 @@
+import { renderToDOM } from './renderToDom';
+import sortStudent from './sortStudent';
+
+// add form to DOM on start-sorting click.
+// Add events for form after the form is on the DOM
+const form = () => {
+  const domString = `<form id="sorting" class="d-flex flex-column form-floating ">
+    <input
+    type="text"
+    class="form-control mb-1"
+    id="student-name"
+    placeholder="Enter a name"
+    required
+  />
+  <label for="floatingInputValue">Name to be sorted</label>
+<button type="submit" class="btn btn-success" id="floatingInputValue">Get Sorted!</button>
+</form>`;
+
+  renderToDOM('#form-container', domString);
+
+  // has to be put on the DOM after form is on DOM, not before
+  // on form submit, sort student
+  document.querySelector('#sorting').addEventListener('submit', sortStudent);
+};
+
+const filterBtnRow = () => {
+  const domString = `<div class="btn-group" role="group" aria-label="Basic example">
+      <button type="button" id="filter--hufflepuff" class="btn btn-warning btn-sm">Hufflepuff</button>
+      <button type="button" class="btn btn-primary btn-sm" id="filter--ravenclaw">Ravenclaw</button>
+      <button type="button" class="btn btn-success btn-sm" id="filter--slytherin">Slytherin</button>
+      <button type="button" class="btn btn-danger btn-sm" id="filter--gryffindor">Gryffindor</button>
+      <button type="button" class="btn btn-secondary btn-sm" id="filter--all">All</button>
+    </div>`;
+
+  renderToDOM('#filter-container', domString);
+};
+
+const studentAreas = () => {
+  const domString = `<div id="students">No Students</div>
+  <div id="voldy">No Death Eaters</div>`;
+
+  renderToDOM('#student-container', domString);
+};
+
+// ********** HTML Components  ********** //
+// the basic HMTL structure of app
+const htmlStructure = () => {
+  const domString = `
+    <div id="header-container" class="header mb-3"></div>
+    <div id="form-container" class="container mb-3 text-center"></div>
+    <div id="filter-container" class="container mb-3"></div>
+    <div id="student-container" class="container d-flex"></div>
+    `;
+
+  renderToDOM('#app', domString);
+};
+
+const header = () => {
+  const domString = `<div class="container">
+    <h1>Welcome to Hoggy Hogwarts Sorting Hat!</h1>
+    <p>
+      Hmm, difficult. VERY difficult. <br />Plenty of courage, I see.
+      <br />Not a bad mind, either. There's talent, oh yes. <br />And a
+      thirst to prove yourself. <br />But where to put you?
+    </p>
+  </div>`;
+
+  renderToDOM('#header-container', domString);
+};
+
+const startSortingBtn = () => {
+  const domString = '<button type="button" class="btn btn-info" id="start-sorting">Start the Sorting Ceremony!</button>';
+
+  renderToDOM('#form-container', domString);
+};
+
+export {
+  form, filterBtnRow, studentAreas, htmlStructure, header, startSortingBtn
+};
