+++ utils/sample_data/students.js
@@ -1,0 +1,3 @@
+const students = [];
+
+export default students;
