+++ utils/sample_data/voldysArmy.js
@@ -1,0 +1,3 @@
+const voldysArmy = []; // starts as an empty array
+
+export default voldysArmy;
